requirements.txt
streamlit>=1.37
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
sentence-transformers>=2.7.0
pdfplumber>=0.11
docx2txt>=0.8
striprtf>=0.0.26
openpyxl>=3.1
python-pptx>=0.6.23
PyPDF2>=3.0.1
beautifulsoup4>=4.12
contractions>=0.1.73
nltk>=3.8
altair>=5.2
matplotlib>=3.8